application/backend/pyproject.toml
[project]
name = "geti-tune"
version = "0.1.0"
description = "Geti Tune"
readme = "readme.md"
requires-python = ">=3.13,<3.14"

dependencies = [
    "fastapi[standard]~=0.121.2",
    "hiyapyco==0.7.0",
    "openvino-model-api~=0.3.0.4",
    "pillow~=12.0",
    "python-multipart~=0.0.20",
    "uvicorn~=0.38.0",
    "psutil~=7.1.3",
    "pyyaml~=6.0.3",
    "openvino==2025.3.0",
    "watchdog==6.0.0",
    "sqlalchemy~=2.0.44",
    "alembic~=1.17.2",
    "aiofiles==25.1.0",
    "pydantic-settings~=2.12.0",
    "aiortc~=1.14.0",
    "datumaro[experimental] @ git+https://github.com/open-edge-platform/datumaro.git@develop",
    "scikit-learn~=1.7.2",
    "faster-coco-eval~=1.7.0",
    "scikit-multilearn~=0.2.0",
    "loguru~=0.7.3",
    "sse-starlette~=3.0.3",
    "cv2-enumerate-cameras==1.3.1",
]

[project.optional-dependencies]
# Choose one of the following training backends:
# - cpu: only CPU training (no GPU support)
# - xpu: support training on compatible Intel GPUs
# - cuda: support training on compatible NVIDIA GPUs
cpu = ["otx[cpu]"]
xpu = ["otx[xpu]"]
cuda = ["otx[cuda]"]

mqtt = [
    "paho-mqtt~=2.1.0",
]
installer = ["pyinstaller~=6.16.0"]

[dependency-groups]
dev = [
    "pytest~=8.4",
    "pytest-asyncio~=1.1.0",
    "testcontainers~=4.10.0",
    "time-machine~=2.19.0",
]
lint = [
    "mypy~=1.16.1",
    "ruff==0.11.13",
    "import-linter==2.3",
    "types-aiofiles~=25.1.0",  # should match aiofiles version
    "types-requests~=2.32.4",
    "types-pyyaml~=6.0.12",  # should match pyyaml version
]

[tool.uv]
default-groups = ["dev", "lint"]
required-version = "~=0.9.7"
conflicts = [
    [
        { extra = "cpu" },
        { extra = "xpu" },
        { extra = "cuda" },
    ],
    # redeclare library conflicts to workaround bug with uv dependency resolution
    # For details, see: https://github.com/astral-sh/uv/issues/9942
    [
        { package = "otx", extra = "cpu" },
        { package = "otx", extra = "xpu" },
        { package = "otx", extra = "cuda" },
    ],
]

[tool.uv.build-backend]
module-root = ""
module-name = "app"

[tool.uv.sources]
otx = { path = "../../library", editable = true }

[tool.ruff]
target-version = "py313"
line-length = 120
exclude = [".venv*"]

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = ["ARG", "E", "F", "I", "N", "UP", "YTT", "ASYNC", "S", "COM", "C4", "FA", "PIE", "PYI", "Q", "RSE", "RET", "SIM",
    "TID", "PL", "RUF", "C90", "D103", "ANN001", "ANN201", "ANN205"]
ignore = ["N801", "N805","N806","N807", "N818", "COM812", "RET503", "SIM108", "SIM105", "PLR2004",
    "RUF010", "RUF012", "CPY001", "PLE1205"]
fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"*test*.py" = ["S", "ANN", "D", "ARG", "PLR", "ASYNC"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pylint]
max-args=7

[tool.importlinter]
root_package = "app"
include_external_packages = true
exclude_type_checking_imports = true

[[tool.importlinter.contracts]]
name = "App Main Vertical"
type = "layers"
id = "main_app"
layers = [
    "app.api",
    "app.services",
    "app.repositories",
    "app.db",
]

[[tool.importlinter.contracts]]
name = "Internal API"
type = "layers"
id = "api"
layers = [
    "app.api.routers",
    "app.api.serializers",
    "app.api.schemas",
    "app.api.dependencies",
]

[[tool.importlinter.contracts]]
name = "Core Module"
type = "layers"
id = "core_module"
layers = [
    "app.core.jobs",
    "app.core.run",
    "app.core.models",
    "app.core.logging",
]

[[tool.importlinter.contracts]]
name = "Service-Utils"
type = "layers"
id = "service_utils"
layers = [
    "app.services",
    "app.utils",
]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
show_error_codes = true
disable_error_code = ["prop-decorator"] # Ignore error due to https://github.com/python/mypy/issues/14461

[tool.pytest.ini_options]
pythonpath = [
    "."
]
